// ProjectFoodAsia/src/app/models/product.ts
export class Product {
    merchantId: number;
    merchantName: string;
    productName: string;
    duration: number;
    savings: number;
    additionalOffers: string;
    imageURL: any;
    merchantDesc: string;
    productDesc: string;
    termsAndconditions: string;
    benefitsAndcancellation: string;
    productType: string;
    taxes: number;
    appFee: number;
    redemption: number;
    minDu: number;
    maxDu: number;
    merchantLogo: any;
    index: any;
    pay: number;
}




export class Grocery{
    merchantId: number;
    merchantName: string;
    productName: string;
    duration: number;
    savings: number;
    additionalOffers: string;
    imageURL: any;
    merchantDesc: string;
    productDesc: string;
    termsAndconditions: string;
    benefitsAndcancellation: string;
    productType: string;
    taxes: number;
    appFee: number;
    redemption: number;
    minDu: number;
    maxDu: number;
    merchantLogo: any;
    index: any;
    pay: number;
    
}


// export class Product {
//     prodId: number;
//     merchantId: number;
//     merchantName: string;
//     productName: string;
//     prodDesc: string;
//     prodImgURL: string; 
//     prodMinDuration: number;
//     prodMaxDuration: number;    
//     prodMinAllowedAmount: number;
//     prodMaxAllowedAmount: number;    
//     prodEffecDate: Date;
//     prodExpiryDate:Date;
//     additionalOffers: string;
//     savings: number;
//     duration: number;
// }